add alias for test
