+++ packages/sdk/tsconfig.json
@@ -1,4 +1,11 @@
 {
 	"extends": "../../tsconfig.json",
+	"compilerOptions": {
+		"baseUrl": "./src",
+		"paths": {
+			"@test/*": ["./test/*"],
+			"@test": ["./test/index.ts"]
+		}
+	},
 	"include": ["src", "tsup.config.ts"]
 }
