+++ packages/sdk/vitest.config.js
@@ -1,11 +1,11 @@
-import { defineConfig } from 'vitest/config';
 import { vanillaExtractPlugin } from '@vanilla-extract/vite-plugin';
+import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
 	plugins: [vanillaExtractPlugin()],
 	test: {
 		environment: 'jsdom',
-		setupFiles: ['./src/react/_internal/test/setup.ts'],
+		setupFiles: ['./test/setup.ts'],
 		include: ['./**/*.test.{ts,tsx}'],
 	},
 });
